requirements.txt
tzdata==2022.5
mysqlclient==2.1.1
scikit-learn==0.24.2
joblib==1.1.0
django-rest-auth
drf-yasg
sqlparse==0.4.3
pytz==2022.5
PyJWT==2.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.0.0
Django==4.1.2
django-cors-headers==3.11.0
asgiref==3.5.2
django-heroku==0.3.1
gunicorn
django-cors-headers
pandas